+++ android/CMakeLists.txt
@@ -16,7 +16,7 @@ set(LibRoot "${CMAKE_CURRENT_SOURCE_DIR}/src/main/jniLibs")
 set(ArchivePath "${CMAKE_CURRENT_SOURCE_DIR}/${LibraryVersion}.tar.gz")
 if(NOT EXISTS ${ArchivePath})
   file(DOWNLOAD
-    "https://github.com/breez/breez-liquid-sdk-flutter/releases/download/${TagName}/android.tar.gz"
+    "https://github.com/breez/breez-sdk-liquid-flutter/releases/download/${TagName}/android.tar.gz"
     ${ArchivePath}
     TLS_VERIFY ON
   )
