+++ ios/Classes/flutter_breez_liquid.c
@@ -1,4 +1,4 @@
-#include "breez_liquid_sdk.h"
+#include "breez_sdk_liquid.h"
 
 int64_t dummy_bundler(void) {
   return dummy_method_to_enforce_bundling();
