+++ pubspec.yaml
@@ -2,7 +2,7 @@ name: flutter_breez_liquid
 description: Flutter wrapper around Dart bindings for the Breez Liquid SDK
 version: 0.1.2-dev5
 homepage: https://breez.technology
-repository: https://github.com/breez/breez-liquid-sdk-flutter
+repository: https://github.com/breez/breez-sdk-liquid-flutter
 publish_to: 'none'
 
 environment:
@@ -14,7 +14,7 @@ dependencies:
     sdk: flutter
   breez_liquid:
     git:
-      url: https://github.com/breez/breez-liquid-sdk-dart
+      url: https://github.com/breez/breez-sdk-liquid-dart
   ffigen: ">=8.0.0 <=11.0.0"
 
 dev_dependencies:
